# src/main/java/io/swagger/api/OrderApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.46).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Order;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-09-28T16:59:58.194634916Z[GMT]")
@Validated
public interface OrderApi {

    @Operation(summary = "Order löschen", description = "Order löschen", tags={ "order" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "400", description = "Ungültige ID"),
        
        @ApiResponse(responseCode = "404", description = "Order nicht gefunden") })
    @RequestMapping(value = "/order/{order_id}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteOrder(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("orderId") Long orderId);


    @Operation(summary = "Einkauf finden", description = "Einkauf finden", tags={ "order" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Order.class))),
        
        @ApiResponse(responseCode = "400", description = "Ungültige ID"),
        
        @ApiResponse(responseCode = "404", description = "Order nicht gefunden") })
    @RequestMapping(value = "/order/{order_id}",
        produces = { "application/json", "application/xml" }, 
        method = RequestMethod.GET)
    ResponseEntity<Order> getOrderById(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("orderId") Long orderId);


    @Operation(summary = "Einkauf abschließen", description = "", tags={ "order" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Einkauf erfolgreich abgeschlossen"),
        
        @ApiResponse(responseCode = "401", description = "Nicht authentifiziert"),
        
        @ApiResponse(responseCode = "403", description = "Nicht autorisiert") })
    @RequestMapping(value = "/order/checkout",
        method = RequestMethod.POST)
    ResponseEntity<Void> orderCheckoutPost();


    @Operation(summary = "Order erstellen", description = "", tags={ "order" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Order erfolgreich erstellt"),
        
        @ApiResponse(responseCode = "400", description = "Ungültige Anforderung"),
        
        @ApiResponse(responseCode = "401", description = "Nicht authentifiziert"),
        
        @ApiResponse(responseCode = "403", description = "Nicht autorisiert") })
    @RequestMapping(value = "/order",
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> orderPost(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody Order body);

}