# src/main/java/io/swagger/api/OrderApiController.java
package io.swagger.api;

import io.swagger.model.Order;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-09-28T16:59:58.194634916Z[GMT]")
@RestController
public class OrderApiController implements OrderApi {

    private static final Logger log = LoggerFactory.getLogger(OrderApiController.class);

    private final ObjectMapper objectMapper;

    private final HttpServletRequest request;

    @org.springframework.beans.factory.annotation.Autowired
    public OrderApiController(ObjectMapper objectMapper, HttpServletRequest request) {
        this.objectMapper = objectMapper;
        this.request = request;
    }

    public ResponseEntity<Void> deleteOrder(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("orderId") Long orderId) {
        String accept = request.getHeader("Accept");
        return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Order> getOrderById(@Parameter(in = ParameterIn.PATH, description = "", required=true, schema=@Schema()) @PathVariable("orderId") Long orderId) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/json")) {
            try {
                return new ResponseEntity<Order>(objectMapper.readValue("{\n  \"shipDate\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"complete\" : true,\n  \"order_id\" : 0,\n  \"user\" : [ {\n    \"password\" : \"password\",\n    \"address\" : [ {\n      \"zip\" : \"94301\",\n      \"city\" : \"Palo Alto\",\n      \"street\" : \"437 Lytton\",\n      \"address_id\" : 0,\n      \"state\" : \"CA\"\n    }, {\n      \"zip\" : \"94301\",\n      \"city\" : \"Palo Alto\",\n      \"street\" : \"437 Lytton\",\n      \"address_id\" : 0,\n      \"state\" : \"CA\"\n    } ],\n    \"id\" : 100000,\n    \"email\" : \"email\",\n    \"username\" : \"fehguy\"\n  }, {\n    \"password\" : \"password\",\n    \"address\" : [ {\n      \"zip\" : \"94301\",\n      \"city\" : \"Palo Alto\",\n      \"street\" : \"437 Lytton\",\n      \"address_id\" : 0,\n      \"state\" : \"CA\"\n    }, {\n      \"zip\" : \"94301\",\n      \"city\" : \"Palo Alto\",\n      \"street\" : \"437 Lytton\",\n      \"address_id\" : 0,\n      \"state\" : \"CA\"\n    } ],\n    \"id\" : 100000,\n    \"email\" : \"email\",\n    \"username\" : \"fehguy\"\n  } ],\n  \"cart\" : [ {\n    \"cart_id\" : 6,\n    \"total\" : 5,\n    \"quantity\" : 1,\n    \"product_id\" : \"product_id\"\n  }, {\n    \"cart_id\" : 6,\n    \"total\" : 5,\n    \"quantity\" : 1,\n    \"product_id\" : \"product_id\"\n  } ],\n  \"status\" : \"approved\"\n}", Order.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<Order>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<Order>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Void> orderCheckoutPost() {
        String accept = request.getHeader("Accept");
        return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Void> orderPost(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody Order body) {
        String accept = request.getHeader("Accept");
        return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
    }

}