# src/main/java/io/swagger/api/ProductApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.46).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import java.math.BigDecimal;
import io.swagger.model.Product;
import java.util.UUID;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-09-28T16:59:58.194634916Z[GMT]")
@Validated
public interface ProductApi {

    @Operation(summary = "Neues Produkt hinzufügen", description = "Neues Produkt hinzufügen", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Erfolgreich hinzufügt", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))),
        
        @ApiResponse(responseCode = "405", description = "Ungültiger Input") })
    @RequestMapping(value = "/product",
        produces = { "application/json", "application/xml" }, 
        consumes = { "application/json", "application/xml", "application/x-www-form-urlencoded" }, 
        method = RequestMethod.POST)
    ResponseEntity<Product> addProduct(@Parameter(in = ParameterIn.DEFAULT, description = "Neues Produkt hinzufügen", required=true, schema=@Schema()) @Valid @RequestBody Product body);


    @Operation(summary = "Produkt löschen", description = "", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "400", description = "Ungültige ID") })
    @RequestMapping(value = "/product/{product_id}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteProduct(@Parameter(in = ParameterIn.PATH, description = "Product id to delete", required=true, schema=@Schema()) @PathVariable("product_id") Long productId);


    @Operation(summary = "Alle Produkte finden", description = "Gibt ein Array mit allen Produkten aus", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Erfolgreich gefunden", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))) })
    @RequestMapping(value = "/product",
        produces = { "application/json", "application/xml" }, 
        method = RequestMethod.GET)
    ResponseEntity<Product> findAllProducts();


    @Operation(summary = "Produkt finden", description = "", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Erfolgreich gefunden", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))),
        
        @ApiResponse(responseCode = "400", description = "Ungültige ID"),
        
        @ApiResponse(responseCode = "404", description = "Produkt nicht gefunden") })
    @RequestMapping(value = "/product/{product_id}",
        produces = { "application/json", "application/xml" }, 
        method = RequestMethod.GET)
    ResponseEntity<Product> findProduct();


    @Operation(summary = "Produkt updaten", description = "Produkt updaten", tags={ "product" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Update erfolgreich", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))),
        
        @ApiResponse(responseCode = "400", description = "Ungültige ID"),
        
        @ApiResponse(responseCode = "404", description = "Produkt nicht gefunden"),
        
        @ApiResponse(responseCode = "405", description = "Validation exception") })
    @RequestMapping(value = "/product/{product_id}",
        produces = { "application/json", "application/xml" }, 
        consumes = { "application/json", "application/xml", "application/x-www-form-urlencoded" }, 
        method = RequestMethod.PUT)
    ResponseEntity<Product> updateProduct(@Parameter(in = ParameterIn.DEFAULT, description = "Produkt updaten", required=true, schema=@Schema()) @Valid @RequestBody Product body);

}